we are done  with the second problem, at the begininng we need two variables, one for the player call and other for the computer's.--> the variables are created after the function to store the arguments we need to call the function. our function needs to declare a winner and a loser based on a series of logical conditions. rock beats scissors, scissors beats paper, paper, beats rock variables playerSelection type string init val '' computerSelection type string init val '' output who win and why. input: player with a prompt function and computer with a function called computerPlay ****IN WHICH CASES PLAYER WINS**** Chooses paper and cpu chooses rock. chooses rock cpu chooses scissors. scissors vs paper
***IN WHICH CASES CPU WINS***
player rock cpu paper. player paper cpu scissors. player scissors cpu rock.

*** IN WHICH CASES THERE IS NO WINNER****
any other combination ends in a tie.

I did not had problem writing the logic conditions to generate the results, but I had problem with the variables creation and the fact that the const to contain the choices had the same name as the arguments kind of confuse me too.